refactor: Extract cart table into CartTable component
